Rework Speed Bonuses and Penalties to Tags
All status bonuses to speed are now using a tag system similar to others stats. Same for status penalties and circumstance penalties. Presently, no other types of speed bonus are needed, particularly as Armour penalties and ancestry speed boosts are typeless. The main benefit of this is that as more submods add spells that effect speeds, it would otherwise be quite easy to get a monk to far higher speeds than intended.