Shield Damage
Shields now track their hit points instead of damage. Shields can be repaired by making a Crafting (Nature) check against a DC according to their rarity at an anvil (which subtracts gold equal to half the base cost). Doing so will make the character walk up to the anvil and do a (placeholder) animation. On a success, 5 + 5 per skill rank is repaired, doubled on a crit. On a critical failure, 2d6 damage is dealt instead. Shields become unequipable at their BT, and destroyed at 0 HP (so there is a consequence to trying to repair a shield). Also turns expertise into a passive, as this was necessary to allow us to detect it. Updates classes to give passive selectors for expertise instead of base game expertise selector.