Scroll Crafting
Magical crafting feat gives a once per day shout spell. Casting it spawns an enemy (without starting combat) with an aura near to the caster. The next spell for which a scroll exists cast in the aura is turned into a scroll if the crafter has sufficient gold and passes a check, after which the enemy disappears. The enemy also disappears after 60 seconds, if the crafter moves, or if another player casts the crafting shout spell. The enemy exists so that spells that require an enemy can be targeted. The caster's casting gets blocked, but they still get charged a spell slot as appropriate. Using a scroll in the aura blocks you from being the caster, to avoid scroll duplication. Blocked characters unlock a shout to unblock themselves.