Spell trait framework
Add condition helpers that can be tested for with HasStringInSpellConditions to mark traits such as Fear, Emotion, Mental, Auditory, Visual, and Concentrate. These conditions also perform the checks that you would expect of them. Add these conditions to the target conditions of a spell. Concentrate is a special case where it should also be added to requirement conditions for a spell. Documentation will be added to the wiki.